Adjusted section 1 and 2 of landing page
In section one the code worked but the way the containers were used was obscure so
instead a header container was made for the top part of section 1 and a content
container was made for the rest of section 1. For section 2 there was a container for
each set of images and text but that was switched with just one container. Also added
the ability to scroll which wasnt initially possible.